removed cmd/registry file
