+++ cmd/epp/main.go
@@ -25,8 +25,6 @@ import (
 )
 
 func main() {
-	// Register all known plugin factories
-	runner.RegisterAllPlugins()
 	// For adding out-of-tree plugins to the plugins registry, use the following:
 	// plugins.Register(my-out-of-tree-plugin-name, my-out-of-tree-plugin-factory-function)
